requirements.txt
pandas
flask
matplotlib
seaborn
numpy
vaderSentiment
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytz
wordcloud
google-generativeai
nltk
